refactor(composio): remove legacy helpers usage; add composio-utils and migrate routes to ComposioClient [Droid-assisted]
