+++ app/api/composio/connect/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from 'next/server'
 import { getServerSession } from 'next-auth'
 import { z } from 'zod'
 import { getAuthOptions } from '@/lib/auth'
-import { isValidToolkit, generateComposioUserId } from '@/lib/composio'
+import { isValidToolkit, generateComposioUserId } from '@/lib/composio-utils'
 import { getWorkspaceWithPermissions } from '@/lib/db/queries'
 import { aiConfig } from '@/lib/env'
 import { zId } from '@/lib/utils/ids'
